Maven Central Upload works now.
Moved the signing closure into the build script root. Before, it was
only exectured during the uploadArchives task and somehow this made the
upload of the archive .asc files fail. Or the signing at all. Don't know
exactly, but now it's working. However, of course, now the signing
always happens which may be troublesome for someone who just wants to
build to plugin but not upload it.