Cleaning. Cannot really avoid the connection test errors.
I tried a lot, including trying to run the original ElasticSearch
tests for the `analyzer-icu` plugin with the hope to learn something.
However, it is not so easy to just run those tests and get their
output. I copied the plugin project in to a Maven project and let
it run there to see all output. Sometimes the connection errors were
there, sometimes not. I will now stop trying to avoid those since
the tests themselves are working and the connection errors are no
test issues but framework issues (they also happen with a test that
does nothing).