Remove the `doXContentBody` method.
The exact purpose of the method is not clear to me since there seems to be a default for creating the XContent of a field mapper in the `FieldMapper` class that includes all the parameters. However, there is a comment about BWC, "probably burrows wheeler" compression.
The issue is, when we leave the method as copied from `TextFieldMapper` that we access the `FieldMapper.Parameter#toXContent` method which is protected in `Parameter`. This seems legal in the IDE since the `PreanalyzedMapper` extends from `FieldMapper`. But in deployment, the plugins reside in their own JARs and are loaded with a different class loader than the ElasticSearch core classes. This leads to an `IllegalAccessError` as soon as a mapping with a preanalyzed field is put onto an index.